Get everything working

+++ publish.sh
@@ -1,14 +1,25 @@
-echo "//registry.npmjs.org/:_authToken=$NPM_TOKEN" > .npmrc
+cd reactfire/pub/reactfire
 
-cp ../README.md .
-cp ../LICENSE .
+if test $NPM_TOKEN; then
 
-if test $TAG_NAME; then
-    npm version $(echo $TAG_NAME | sed 's/^v\(.*\)$/\1/')
-    npm publish . --tag next
-else
-    npm version $(npm version | head -n 1 |  sed "s/^.*: '\([^']*\).*/\1/")-canary.$SHORT_SHA
-    npm publish . --tag canary
+    echo "//registry.npmjs.org/:_authToken=$NPM_TOKEN" > .npmrc
+
+    cp ../../../README.md .
+    cp -r ../../../docs .
+    cp ../../../LICENSE .
+
+    if test $TAG_NAME; then
+        npm version $(echo $TAG_NAME | sed 's/^v\(.*\)$/\1/')
+        npm publish . --tag next
+        ret=$?
+    else
+        npm version $(npm version | sed -n "s/^  reactfire: '\(.*\)',/\1/p")-canary.$SHORT_SHA
+        npm publish . --tag canary
+        ret=$?
+    fi
+
+    rm -f .npmrc
+
 fi
 
-rm -f .npmrc
+exit $ret
